backend/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
pyjwt
passlib
bcrypt
